backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
motor>=3.6.0
pymongo>=4.11.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pydantic[email]>=2.11.0
pydantic-settings>=2.7.0
python-multipart>=0.0.19
